+++ apps/expo-app/app/(tabs)/story/_layout.tsx
@@ -1,7 +1,9 @@
 import { Stack } from 'expo-router';
 
 const StoryLayout = () => {
-  return <Stack screenOptions={{ headerShown: false }} />;
+  return (
+    <Stack screenOptions={{ headerShown: false, gestureEnabled: false }} />
+  );
 };
 
 export default StoryLayout;
