+++ apps/expo-app/app/_layout.tsx
@@ -4,7 +4,11 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 const RootLayout = () => {
   return (
     <GestureHandlerRootView>
-      <Stack>
+      <Stack
+        screenOptions={{
+          gestureEnabled: false,
+        }}
+      >
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
         <Stack.Screen name="(auth)" options={{ headerShown: false }} />
       </Stack>
